Modificaciones que hizo Richelle en el proyecto

+++ src/modelo/Prestamo.java
@@ -4,7 +4,6 @@
  * 
  */
 package modelo;
-
 /**
  *
  * @author Canela
@@ -14,9 +13,9 @@ public class Prestamo {
     private String nombre;
     private String direccion;
     private String noTelefono;
-    private double importe;
+    private int importe;
 
-    public Prestamo(int noPrestamo, String nombre, String direccion, String noTelefono, double importe) {
+    public Prestamo(int noPrestamo, String nombre, String direccion, String noTelefono, int importe) {
         this.noPrestamo = noPrestamo;
         this.nombre = nombre;
         this.direccion = direccion;
@@ -40,13 +39,12 @@ public class Prestamo {
         return noTelefono;
     }
 
-    public double getImporte() {
+    public int getImporte() {
         return importe;
     }
 
     @Override
     public String toString() {
         return "Numero de prestamo: " + noPrestamo + "\nNombre del cliente: " + nombre + "\nDireccion: " + direccion + "\nTelefono: " + noTelefono + "\nImporte solicitado: " + importe + '\n';
-    }
-    
+    } 
 }
